DeepSpace-Attitude-Lab/requirements.txt
streamlit==1.39.0
numpy==1.26.4
scipy==1.13.1
matplotlib==3.8.4
plotly==5.23.0
pillow==10.3.0
kaleido==0.2.1
python-pptx==0.6.23
